F_T32247 move ReportBundle Logic to core
- adjust use statements
