+++ demosplan/DemosPlanUserBundle/Logic/MasterToebService.php
@@ -19,8 +19,8 @@ use demosplan\DemosPlanCoreBundle\Entity\User\User;
 use demosplan\DemosPlanCoreBundle\Logic\CoreService;
 use demosplan\DemosPlanCoreBundle\Logic\DateHelper;
 use demosplan\DemosPlanCoreBundle\Logic\EntityHelper;
-use demosplan\DemosPlanReportBundle\Logic\MasterPublicAgencyReportEntryFactory;
-use demosplan\DemosPlanReportBundle\Logic\ReportService;
+use demosplan\DemosPlanCoreBundle\Logic\Report\MasterPublicAgencyReportEntryFactory;
+use demosplan\DemosPlanCoreBundle\Logic\Report\ReportService;
 use demosplan\DemosPlanUserBundle\Repository\MasterToebRepository;
 use demosplan\DemosPlanUserBundle\Repository\MasterToebVersionRepository;
 use Doctrine\ORM\EntityNotFoundException;
